requirements.txt
annotated-types==0.7.0
anyio==4.11.0
argcomplete==3.6.2
blinker==1.9.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
cryptography==46.0.3
deprecation==2.1.0
Flask==3.1.2
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
multidict==6.7.0
packaging==25.0
pipx==1.8.0
platformdirs==4.5.0
postgrest==2.22.2
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dotenv==1.1.1
realtime==2.22.2
requests==2.32.5
setuptools==80.9.0
sniffio==1.3.1
storage3==2.22.2
StrEnum==0.4.15
supabase==2.22.2
supabase-auth==2.22.2
supabase-functions==2.22.2
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
userpath==1.9.2
websockets==15.0.1
Werkzeug==3.1.3
wheel==0.45.1
yarl==1.22.0
